Refactor of data access.
