+++ pipe/location_active_dates/location_active_period_loader.json
@@ -4,7 +4,7 @@
     },
     "transform": {
         "image_pull_secrets": [ "battelleecology-quay-read-all-pull-secret" ],
-        "image": "quay.io/battelleecology/location_active_dates:20",
+        "image": "quay.io/battelleecology/location_active_dates:21",
         "cmd": [ "/bin/bash" ],
         "stdin": [ "#!/bin/bash", "python3 -m location_active_dates.active_period_loader" ],
         "env": {
